chore:#29 충돌문제 해결
