+++ src/main/java/com/back/cafe/domain/product/dto/ProductRequestDto/ProductCreateRequest.java
@@ -1,9 +1,12 @@
-package com.back.cafe.domain.product.dto;
+package com.back.cafe.domain.product.dto.ProductRequestDto;
 
+import jakarta.validation.constraints.NotBlank;
+
+import com.back.cafe.domain.product.dto.ServiceDto.ServiceCreateProductDto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record ProductModifyRequest(
+public record ProductCreateRequest(
         @NotBlank String name,
         @NotBlank String category,
         @NotNull Long price,
@@ -11,8 +14,8 @@ public record ProductModifyRequest(
         @NotBlank String description,
         @NotBlank String imageUrl
 ) {
-    public ServiceModifyProductDto toServiceDto(){
-        return new ServiceModifyProductDto(
+    public ServiceCreateProductDto toServiceDto(){
+        return new ServiceCreateProductDto(
                 name,
                 category,
                 price,
@@ -21,6 +24,4 @@ public record ProductModifyRequest(
                 imageUrl
         );
     }
-
 }
-
